Add benchmarks
Add some benchmarks to make sure that there's some benefit to this
project. Preliminiarily, it looks like theres a 33% increase. No heavy
I/O is implemented so far, so this may change. But the increase in
performance is reassuring that this is the right path, instead of
writing a wrapper around the SWIG bindings.